Section 2: Thread expert_counts through CUDA path for expert recycling
- Change expert_counts buffer from torch.long to torch.int32 (matches CUDA int*)
- Add expert_counts parameter to launch_mamba3_peer_step and
  launch_mamba3_peer_batched_step (ops.h, ops.cpp, kernels, pybind)
- Replace nullptr with expert_counts.data_ptr<int>() in both launchers
- Add step_counter increment and periodic _recycle_dead_experts call
  after CUDA batched step in supergrok2.py

+++ grokking_race_package/grokking_optimizers/grokking_optimizers/mamba3_peer_metanet.py
@@ -263,7 +263,7 @@ class Mamba3PEERMetaNet(nn.Module):
         self.expert_b2 = nn.Parameter(torch.zeros(num_experts, 1))
 
         # -- Expert activation tracking (not a parameter)
-        self.register_buffer('expert_counts', torch.zeros(num_experts, dtype=torch.long))
+        self.register_buffer('expert_counts', torch.zeros(num_experts, dtype=torch.int32))
         self.register_buffer('step_counter', torch.tensor(0, dtype=torch.long))
 
     def forward(
